requirements.txt
absl-py==0.12.0
addcopyfighandler==3.0.0
aiohttp==3.7.1
aiohttp-cors==0.7.0
aioredis==1.3.1
altair==4.1.0
ansi2html==1.6.0
anyio==2.2.0
appdirs==1.4.4
argon2-cffi==20.1.0
asciitree==0.3.3
astor==0.8.1
asttokens==2.0.4
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
Authlib==0.15.5
Automat==20.2.0
Babel==2.9.0
backcall==0.2.0
bankroll==0.4.1
bankroll-broker==0.4.2
bankroll-broker-fidelity==0.4.0
bankroll-marketdata==0.4.0
bankroll-model==0.4.1
beautifulsoup4==4.9.1
blackstone==0.1.15
bleach==3.2.1
blessings==1.7
blis==0.7.4
blockdiag==2.0.1
bokeh==2.4.0
bokeh-wordcloud2==0.9.1
boto3==1.16.30
botocore==1.19.30
branca==0.4.1
brewer2mpl==1.4.1
Brotli==1.0.9
CacheControl==0.12.6
cachelib==0.1.1
cachetools==4.2.1
camelot-py==0.10.1
catalogue==2.0.1
certifi==2020.6.20
cffi==1.14.3
cftime==1.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.4
colorcet==2.0.6
colorful==0.5.4
conllu==4.4
constantly==15.1.0
cramjam==2.4.0
cryptography==3.3.1
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.5
Cython==0.29.14
dash==1.19.0
dash-bootstrap-components==0.12.0
dash-colorscales==0.0.4
dash-core-components==1.15.0
dash-daq==0.5.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
dask==2021.2.0
datasets==1.6.2
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
distlib==0.3.1
distro==1.5.0
dlx==1.0.4
docplex==2.20.204
docx-utils==0.1.3
dtale==1.41.1
empyrical==0.5.5
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.0.0/en_core_web_md-3.0.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
entrypoints==0.3
et-xmlfile==1.0.1
executing==0.5.4
facebook-sdk==3.1.0
fastdtw==0.3.4
fasteners==0.16.3
fastjsonschema==2.15.1
fastparquet==0.7.1
filelock==3.0.12
firebase-admin==5.2.0
flashtext==2.7
Flask==1.1.2
Flask-Compress==1.9.0
Flask-Cors==3.0.10
flask-ngrok==0.0.25
Flask-OAuthlib==0.9.6
flatbuffers==1.12
fsspec==0.8.7
funcparserlib==0.3.6
future==0.18.2
gast==0.3.3
gensim==3.8.3
geographiclib==1.50
geopy==2.0.0
ggplot==0.11.5
ghostscript==0.7
gitdb==4.0.7
GitPython==3.1.14
google==3.0.0
google-api-core==1.26.0
google-api-python-client==2.35.0
google-auth==1.27.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.41.1
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
googlemaps==4.4.2
gpustat==0.6.0
graphviz==0.16
grpcio==1.32.0
h5py==2.10.0
hiredis==1.1.0
holoviews==1.14.7
httplib2==0.18.1
huggingface-hub==0.0.8
hvplot==0.7.3
hyperlink==20.0.1
icecream==2.1.0
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
inflection==0.5.1
iniconfig==1.1.1
intake==0.6.3
intake-parquet==0.2.3
intake-xarray==0.5.0
ipydrawio==1.1.1
ipykernel==5.3.4
ipyleaflet==0.13.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
iso3166==1.0.1
iso8601==0.1.16
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyter-dash==0.4.0
jupyter-packaging==0.7.12
jupyter-server==1.5.1
jupyter-server-mathjax==0.2.2
jupyterlab==3.1.18
jupyterlab-git==0.30.0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.4.0
jupyterlab-widgets==1.0.2
kaleido==0.2.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
locket==0.2.1
logger==1.4
lxml==4.5.2
lz4==3.1.3
mapbox==0.18.0
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.11.1
matplotlib==3.2.2
mistune==0.8.4
mock==4.0.3
modin==0.8.3
more-itertools==8.7.0
mpmath==1.2.1
msgpack==1.0.0
multidict==5.0.0
multiprocess==0.70.11.1
multitasking==0.0.9
murmurhash==1.0.5
nbclassic==0.2.6
nbclient==0.5.1
nbconvert==6.0.7
nbdime==3.0.0
nbformat==5.0.8
nbmultitask==0.1.0
nerodia==0.14.1
nest-asyncio==1.4.2
netCDF4==1.5.7
networkx==2.5
nltk==3.3
nmslib==2.1.1
notebook==6.1.5
noteslib==2.1.0
ntlm-auth==1.5.0
numcodecs==0.9.1
numpy==1.20.2
nvidia-ml-py3==7.352.0
oauthlib==2.1.0
opencensus==0.7.12
opencensus-context==0.1.2
opencv-python==4.5.3.56
openpyxl==3.0.4
opt-einsum==3.3.0
ordereddict==1.1
packaging==20.4
paho-mqtt==1.5.1
pandas==1.3.3
pandas-bokeh==0.5.5
pandas-datareader==0.9.0
pandasgui==0.2.10.1
pandocfilters==1.4.3
panel==0.12.4
param==1.11.1
parsel==1.6.0
parso==0.7.1
partd==1.1.0
pathy==0.4.0
patsy==0.5.1
pdfminer==20191125
pdfminer.six==20201018
pdfplumber==0.5.27
pdftopng==0.2.3
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plac==0.9.6
plotly==4.14.3
pluggy==0.13.1
ply==3.11
polyline==1.4.0
ppscore==1.2.0
preshed==3.0.5
progress==1.5
prometheus-client==0.8.0
prompt-toolkit==3.0.8
Protego==0.1.16
proto-plus==1.19.8
protobuf==3.19.3
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
py-spy==0.3.4
py-trello==0.17.1
pyarrow==1.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.1
pycparser==2.20
pycryptodome==3.10.1
pyct==0.4.8
pydantic==1.7.3
PyDispatcher==2.0.5
pyfolio==0.9.2
Pygments==2.7.2
pygsheets==2.0.5
PyHamcrest==2.0.2
pynput==1.7.3
pyodbc==4.0.30
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
PyQtWebEngine==5.15.4
PyQtWebEngine-Qt5==5.15.2
pyrate-limiter==2.6.0
pyrsistent==0.17.3
pysbd==0.3.4
pysendpulse==0.1.2
pytest==6.2.3
pytextrank==3.1.0
python-constraint==1.4.0
python-dateutil==2.8.1
python-docx==0.8.10
python-edgar==3.1.3
python-pptx==0.6.18
python-sec==0.1.1
python-xbrl==1.1.1
python3-memcached==1.51
pytz==2020.1
pyviz-comms==2.1.0
pywin32==228
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
qiskit==0.26.0
qiskit-aer==0.8.2
qiskit-aqua==0.9.1
qiskit-ibmq-provider==0.13.1
qiskit-ignis==0.6.0
qiskit-terra==0.17.3
qtconsole==4.7.7
QtPy==1.9.0
Quandl==3.6.0
queuelib==1.5.0
ray==1.2.0
redis==3.5.3
regex==2021.3.17
requests==2.24.0
requests-ntlm==1.1.0
requests-oauthlib==1.1.0
requests-toolbelt==0.9.1
retrying==1.3.3
retworkx==0.8.0
robotframework==5.0
rope==0.18.0
rsa==4.7.2
s3transfer==0.3.3
sacremoses==0.0.45
salesforce==0.0.1
scattertext==0.1.2
scikit-learn==0.23.1
scipy==1.5.1
scispacy==0.4.0
Scrapy==2.4.1
seaborn==0.11.1
sec-edgar==0.0.2
sec-edgar-api==1.0.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
sharepoint==0.4.2
sidecar==0.4.0
simple-salesforce==1.11.4
six==1.15.0
sklearn==0.0
slackclient==2.9.3
smart-open==3.0.0
smmap==4.0.0
sniffio==1.2.0
sortedcontainers==2.3.0
soupsieve==2.0.1
spacy==3.0.5
spacy-legacy==3.0.1
spacy-universal-sentence-encoder==0.4.0
spacy-wordnet==0.0.4
SQLAlchemy==1.3.18
sqlalchemy-access==1.0.6
squarify==0.4.3
srsly==2.4.0
statsmodels==0.12.0
strsimpy==0.2.0
superset==0.30.1
swifter==1.0.7
sympy==1.8
tabloo==0.0.13
tabula-py==2.2.0
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
tensorflow-hub==0.11.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
thinc==8.0.2
threadpoolctl==2.1.0
thrift==0.15.0
tinydb==4.6.1
tokenizers==0.10.2
toml==0.10.2
toolz==0.11.1
torch==1.8.1
tornado==6.1
tqdm==4.49.0
traitlets==5.0.5
traittypes==0.2.1
transformers==4.5.1
Twisted==20.3.0
typer==0.3.2
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.25.9
vega==3.4.0
vega-datasets==0.9.0
virtualenv==20.0.26
w3lib==1.22.0
Wand==0.6.6
wasabi==0.8.2
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
websockets==9.0.2
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wikipedia==1.4.0
wiotp-sdk==0.11.0
wordcloud==1.7.0
wrapt==1.12.1
WTForms==2.3.3
xarray==0.17.0
xbrl==0.0.1
xlrd==1.2.0
XlsxWriter==1.2.9
xxhash==2.0.2
yarl==1.6.2
yfinance==0.1.55
zarr==2.10.1
zipp==3.1.0
zope.interface==5.2.0
